src-python/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
langchain-openai>=0.2.5
langgraph>=0.2.45
python-dotenv>=1.0.0
websockets>=13.0
